Fix instantiating NSMutableAttributeString with attributes on Xcode 10's Swift compiler
When instantiating an `NSMutableAttributedString` in a target that imports SwiftyAttributes on Xcode 10 the compiler was getting confused on which initializer to use. The convenience initializer added to `NSAttributedString` seems to be throwing the compiler out of whack. This updates argument names to give a clearer distinction and resolves the issue.

See issue #29 for more details.